feat(web): ultraplan v4.5 — a11y sweep, dynamic titles, CreditChecks refactor
Phase 5C — A11y sweep:
- autoComplete attributes on 12 inputs across 5 pages (name, tel, email, password)

Phase 5D — UX improvements:
- useDocumentTitle hook + applied to 20 pages (dynamic titles for detail pages)
- 13 hardcoded hex colors → CSS custom properties in index.css

Phase 3C — Refactoring:
- Split CreditChecksPage (1382 lines) → 7 files:
  types, Filters, Table, Detail, CreateModal, OverrideModal, index
  (largest file now ~370 lines)

577 API tests (26 suites), TypeScript 0 errors (API + Web).
