+++ apps/web/src/hooks/useDocumentTitle.ts
@@ -1,0 +1,12 @@
+import { useEffect } from 'react';
+
+const BASE_TITLE = 'BESTCHOICE';
+
+export function useDocumentTitle(title?: string) {
+  useEffect(() => {
+    document.title = title ? `${title} | ${BASE_TITLE}` : BASE_TITLE;
+    return () => {
+      document.title = BASE_TITLE;
+    };
+  }, [title]);
+}
