--- apps/web/src/pages/CreditChecksPage.tsx
@@ -1,1382 +1,0 @@
-import { useState, useRef, useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
-import { exportToExcel, type ExcelColumn } from '@/utils/excel.util';
-import api, { getErrorMessage } from '@/lib/api';
-import { compressImageForOcr } from '@/lib/compressImage';
-import { useDebounce } from '@/hooks/useDebounce';
-import { formatDateShort } from '@/utils/formatters';
-import PageHeader from '@/components/ui/PageHeader';
-import DataTable from '@/components/ui/DataTable';
-import QueryBoundary from '@/components/QueryBoundary';
-import Modal from '@/components/ui/Modal';
-import { useAuth } from '@/contexts/AuthContext';
-import { toast } from 'sonner';
-import ThaiDateInput from '@/components/ui/ThaiDateInput';
-import { ArrowLeft, User, FileText, Clock, BarChart3 } from 'lucide-react';
-
-interface OcrBookBankResult {
-  accountName: string | null;
-  accountNo: string | null;
-  bankName: string | null;
-  branchName: string | null;
-  accountType: string | null;
-  balance: number | null;
-  lastTransactionDate: string | null;
-  confidence: number;
-}
-
-interface Customer {
-  id: string;
-  name: string;
-  phone: string;
-  nationalId: string;
-  salary: string | null;
-  occupation: string | null;
-  addressCurrentType: string | null;
-  salaryPayDay: number | null;
-}
-
-interface OcrSalarySlipResult {
-  netSalary: number | null;
-  employerName: string | null;
-  slipDate: string | null;
-  payDay: number | null;
-  bankName: string | null;
-  confidence: number;
-}
-
-interface OcrBankStatementResult {
-  accountName: string | null;
-  bankName: string | null;
-  totalIncome: number | null;
-  totalExpense: number | null;
-  balance: number | null;
-  transactionCount: number | null;
-  dateRange: string | null;
-  confidence: number;
-}
-
-interface CustomerHistory {
-  totalContracts: number;
-  closedContracts: number;
-  activeContracts: number;
-  onTimePaymentPct: number;
-  currentOutstanding: number;
-  isReturning: boolean;
-}
-
-interface RiskScoreResult {
-  riskLevel: 'LOW' | 'MEDIUM' | 'HIGH';
-  score: number;
-  debtToIncome: number;
-  recommendedPayDay: number | null;
-  recommendation: string;
-}
-
-const BANK_OPTIONS = [
-  'กสิกรไทย',
-  'กรุงเทพ',
-  'กรุงไทย',
-  'ไทยพาณิชย์',
-  'ออมสิน',
-  'ธ.ก.ส.',
-  'กรุงศรี',
-  'ทหารไทยธนชาต',
-];
-
-interface AiAnalysisData {
-  monthlyIncome?: number;
-  averageBalance?: number;
-  affordabilityRatio?: number;
-  incomeConsistency?: string;
-  riskFactors?: string[];
-  positiveFactors?: string[];
-  [key: string]: unknown;
-}
-
-interface CreditCheckItem {
-  id: string;
-  status: string;
-  bankName: string | null;
-  statementFiles: string[];
-  statementMonths: number;
-  aiScore: number | null;
-  aiSummary: string | null;
-  aiRecommendation: string | null;
-  aiAnalysis: AiAnalysisData | null;
-  reviewNotes: string | null;
-  checkedBy: { id: string; name: string } | null;
-  customer: { id: string; name: string; phone: string; salary: string | null; occupation: string | null };
-  contract: { id: string; contractNumber: string } | null;
-  createdAt: string;
-}
-
-interface CreditCheckSummary {
-  totalCount: number;
-  pendingCount: number;
-  approvedCount: number;
-  rejectedCount: number;
-  avgScore: number;
-}
-
-interface CreditChecksResponse {
-  data: CreditCheckItem[];
-  total: number;
-  page: number;
-  limit: number;
-  totalPages: number;
-  summary: CreditCheckSummary;
-}
-
-interface Branch {
-  id: string;
-  name: string;
-}
-
-const statusLabels: Record<string, { label: string; className: string }> = {
-  PENDING: { label: 'รอวิเคราะห์', className: 'bg-muted text-foreground' },
-  APPROVED: { label: 'ผ่าน', className: 'bg-success/10 text-success dark:bg-success/15' },
-  REJECTED: { label: 'ไม่ผ่าน', className: 'bg-destructive/10 text-destructive dark:bg-destructive/15' },
-  MANUAL_REVIEW: { label: 'ต้องตรวจเพิ่ม', className: 'bg-amber-100 text-amber-700' },
-};
-
-function getRiskBadge(aiScore: number | null) {
-  if (aiScore === null) {
-    return <span className="px-2 py-0.5 rounded-full text-xs font-medium bg-gray-100 text-gray-500">รอวิเคราะห์</span>;
-  }
-  if (aiScore >= 70) {
-    return <span className="px-2 py-0.5 rounded-full text-xs font-medium bg-success/10 text-success dark:bg-success/15">ความเสี่ยงต่ำ</span>;
-  }
-  if (aiScore >= 50) {
-    return <span className="px-2 py-0.5 rounded-full text-xs font-medium bg-warning/10 text-warning dark:bg-warning/15">ความเสี่ยงปานกลาง</span>;
-  }
-  if (aiScore >= 40) {
-    return <span className="px-2 py-0.5 rounded-full text-xs font-medium bg-warning/10 text-warning dark:bg-warning/15">ต้องตรวจเพิ่ม</span>;
-  }
-  return <span className="px-2 py-0.5 rounded-full text-xs font-medium bg-destructive/10 text-destructive dark:bg-destructive/15">ความเสี่ยงสูง</span>;
-}
-
-export default function CreditChecksPage() {
-  const navigate = useNavigate();
-  const queryClient = useQueryClient();
-  const { user } = useAuth();
-  const isOwner = user?.role === 'OWNER';
-  const isOwnerOrManager = ['OWNER', 'BRANCH_MANAGER'].includes(user?.role ?? '');
-
-  // Filters
-  const [search, setSearch] = useState('');
-  const [statusFilter, setStatusFilter] = useState('');
-  const [startDate, setStartDate] = useState('');
-  const [endDate, setEndDate] = useState('');
-  const [branchFilter, setBranchFilter] = useState('');
-  const [page, setPage] = useState(1);
-  const debouncedSearch = useDebounce(search);
-
-  // Expand row
-  const [expandedRow, setExpandedRow] = useState<string | null>(null);
-
-  // Create full-screen overlay
-  const [showCreateModal, setShowCreateModal] = useState(false);
-  const [customerSearch, setCustomerSearch] = useState('');
-  const debouncedCustomerSearch = useDebounce(customerSearch);
-  const [selectedCustomer, setSelectedCustomer] = useState<Customer | null>(null);
-  const [bankName, setBankName] = useState('');
-  const fileRef = useRef<HTMLInputElement>(null);
-  const bookBankFileRef = useRef<HTMLInputElement>(null);
-  const [bookBankLoading, setBookBankLoading] = useState(false);
-  const [bookBankResult, setBookBankResult] = useState<OcrBookBankResult | null>(null);
-
-  // Section 2: Salary slip OCR
-  const salarySlipFileRef = useRef<HTMLInputElement>(null);
-  const [salarySlipFiles, setSalarySlipFiles] = useState<File[]>([]);
-  const [salarySlipLoading, setSalarySlipLoading] = useState(false);
-  const [salarySlipResult, setSalarySlipResult] = useState<OcrSalarySlipResult | null>(null);
-  const [salarySlipEditable, setSalarySlipEditable] = useState<{ netSalary: string; employerName: string; payDay: string; bankName: string }>({ netSalary: '', employerName: '', payDay: '', bankName: '' });
-
-  // Section 2: Bank statement OCR
-  const [statementBankName, setStatementBankName] = useState('');
-  const statementFileRef = useRef<HTMLInputElement>(null);
-  const [statementFiles, setStatementFiles] = useState<File[]>([]);
-  const [statementLoading, setStatementLoading] = useState(false);
-  const [statementResult, setStatementResult] = useState<OcrBankStatementResult | null>(null);
-
-  // Section 3: Customer history
-  const [customerHistory, setCustomerHistory] = useState<CustomerHistory | null>(null);
-
-  // Section 4: Risk score
-  const [riskScore, setRiskScore] = useState<RiskScoreResult | null>(null);
-  const [riskLoading, setRiskLoading] = useState(false);
-  const [reviewNotesDraft, setReviewNotesDraft] = useState('');
-
-  // Override
-  const canOverride = user && ['OWNER', 'BRANCH_MANAGER'].includes(user.role);
-  const [overrideId, setOverrideId] = useState<string | null>(null);
-  const [overrideCustomerId, setOverrideCustomerId] = useState<string | null>(null);
-  const [overrideStatus, setOverrideStatus] = useState('');
-  const [overrideNotes, setOverrideNotes] = useState('');
-
-  // Reset page when filters change
-  useEffect(() => {
-    setPage(1);
-  }, [debouncedSearch, statusFilter, startDate, endDate, branchFilter]);
-
-  // Date range shortcuts
-  const setDateRange = (type: 'today' | 'week' | 'month') => {
-    const now = new Date();
-    const fmt = (d: Date) => d.toISOString().slice(0, 10);
-    setEndDate(fmt(now));
-    if (type === 'today') {
-      setStartDate(fmt(now));
-    } else if (type === 'week') {
-      const day = now.getDay();
-      const diff = day === 0 ? 6 : day - 1;
-      const monday = new Date(now);
-      monday.setDate(now.getDate() - diff);
-      setStartDate(fmt(monday));
-    } else {
-      const first = new Date(now.getFullYear(), now.getMonth(), 1);
-      setStartDate(fmt(first));
-    }
-  };
-
-  // Build query params helper
-  const buildParams = (overrideLimit?: number) => {
-    const params = new URLSearchParams();
-    if (debouncedSearch) params.set('search', debouncedSearch);
-    if (statusFilter) params.set('status', statusFilter);
-    if (startDate) params.set('startDate', startDate);
-    if (endDate) params.set('endDate', endDate);
-    if (branchFilter) params.set('branchId', branchFilter);
-    params.set('page', String(page));
-    params.set('limit', String(overrideLimit ?? 20));
-    return params;
-  };
-
-  // Main data query
-  const {
-    data: creditChecksData,
-    isLoading,
-    isError,
-    error,
-    refetch,
-  } = useQuery<CreditChecksResponse>({
-    queryKey: ['credit-checks', debouncedSearch, statusFilter, startDate, endDate, branchFilter, page],
-    queryFn: async () => {
-      const { data } = await api.get(`/credit-checks?${buildParams()}`);
-      return data;
-    },
-  });
-
-  const creditChecks = creditChecksData?.data || [];
-  const summary = creditChecksData?.summary;
-
-  // Branches query (OWNER only)
-  const { data: branches = [] } = useQuery<Branch[]>({
-    queryKey: ['branches-list'],
-    queryFn: async () => {
-      const { data } = await api.get('/branches');
-      return Array.isArray(data) ? data : data.data || [];
-    },
-    enabled: !!isOwner,
-  });
-
-  const { data: customers = [] } = useQuery<Customer[]>({
-    queryKey: ['customers-search-cc', debouncedCustomerSearch],
-    queryFn: async () => {
-      const params = new URLSearchParams();
-      if (debouncedCustomerSearch) params.set('search', debouncedCustomerSearch);
-      const { data } = await api.get(`/customers?${params}`);
-      return data.data || [];
-    },
-    enabled: showCreateModal,
-  });
-
-  // Load customer history when customer selected
-  useEffect(() => {
-    if (!selectedCustomer) {
-      setCustomerHistory(null);
-      setRiskScore(null);
-      setSalarySlipResult(null);
-      setSalarySlipFiles([]);
-      setStatementResult(null);
-      setStatementFiles([]);
-      setStatementBankName('');
-      setReviewNotesDraft('');
-      setSalarySlipEditable({ netSalary: '', employerName: '', payDay: '', bankName: '' });
-      return;
-    }
-    api.get(`/credit-checks/customer-history/${selectedCustomer.id}`)
-      .then(({ data }) => setCustomerHistory(data))
-      .catch(() => setCustomerHistory(null));
-  }, [selectedCustomer?.id, selectedCustomer]);
-
-  // Salary slip OCR handler
-  const handleSalarySlipOcr = async () => {
-    if (salarySlipFiles.length === 0) { toast.error('กรุณาเลือกรูปสลิปเงินเดือน'); return; }
-    setSalarySlipLoading(true);
-    try {
-      const imageBase64 = await compressImageForOcr(salarySlipFiles[0]);
-      const { data } = await api.post<OcrSalarySlipResult>('/ocr/salary-slip', { imageBase64 }, { timeout: 90000 });
-      setSalarySlipResult(data);
-      setSalarySlipEditable({
-        netSalary: data.netSalary?.toString() || '',
-        employerName: data.employerName || '',
-        payDay: data.payDay?.toString() || '',
-        bankName: data.bankName || '',
-      });
-      const pct = (data.confidence * 100).toFixed(0);
-      toast.success(`วิเคราะห์สลิปเงินเดือนสำเร็จ (ความมั่นใจ ${pct}%)`);
-    } catch (err) {
-      toast.error(getErrorMessage(err));
-    } finally {
-      setSalarySlipLoading(false);
-    }
-  };
-
-  // Bank statement OCR handler
-  const handleStatementOcr = async () => {
-    if (statementFiles.length === 0) { toast.error('กรุณาเลือกรูป Statement'); return; }
-    setStatementLoading(true);
-    try {
-      const imageBase64 = await compressImageForOcr(statementFiles[0]);
-      const { data } = await api.post<OcrBankStatementResult>('/ocr/bank-statement', { imageBase64 }, { timeout: 90000 });
-      setStatementResult(data);
-      const pct = (data.confidence * 100).toFixed(0);
-      toast.success(`วิเคราะห์ Statement สำเร็จ (ความมั่นใจ ${pct}%)`);
-    } catch (err) {
-      toast.error(getErrorMessage(err));
-    } finally {
-      setStatementLoading(false);
-    }
-  };
-
-  // Risk score calculation
-  const handleCalculateRisk = async (creditCheckId: string) => {
-    setRiskLoading(true);
-    try {
-      const { data } = await api.post<RiskScoreResult>(`/credit-checks/${creditCheckId}/calculate-risk`);
-      setRiskScore(data);
-      toast.success('คำนวณ Risk Score สำเร็จ');
-    } catch (err) {
-      toast.error(getErrorMessage(err));
-    } finally {
-      setRiskLoading(false);
-    }
-  };
-
-  const uploadMutation = useMutation({
-    mutationFn: async (files: FileList) => {
-      if (!selectedCustomer) throw new Error('เลือกลูกค้าก่อน');
-      const fileUrls: string[] = [];
-      for (const file of Array.from(files)) {
-        const reader = new FileReader();
-        const url = await new Promise<string>((resolve, reject) => {
-          reader.onload = () => resolve(reader.result as string);
-          reader.onerror = () => reject(new Error('ไม่สามารถอ่านไฟล์ได้'));
-          reader.readAsDataURL(file);
-        });
-        fileUrls.push(url);
-      }
-      const { data } = await api.post(`/customers/${selectedCustomer.id}/credit-check`, {
-        bankName: bankName || undefined,
-        statementFiles: fileUrls,
-        statementMonths: 3,
-      });
-      return data;
-    },
-    onSuccess: () => {
-      toast.success('สร้างรายการตรวจเครดิตสำเร็จ');
-      queryClient.invalidateQueries({ queryKey: ['credit-checks'] });
-      setShowCreateModal(false);
-      setSelectedCustomer(null);
-      setBankName('');
-      setCustomerSearch('');
-      setBookBankResult(null);
-      setSalarySlipResult(null);
-      setSalarySlipFiles([]);
-      setStatementResult(null);
-      setStatementFiles([]);
-      setStatementBankName('');
-      setReviewNotesDraft('');
-      setRiskScore(null);
-      setSalarySlipEditable({ netSalary: '', employerName: '', payDay: '', bankName: '' });
-      if (fileRef.current) fileRef.current.value = '';
-    },
-    onError: (err: unknown) => toast.error(getErrorMessage(err)),
-  });
-
-  const analyzeMutation = useMutation({
-    mutationFn: async ({ customerId, creditCheckId }: { customerId: string; creditCheckId: string }) => {
-      const { data } = await api.post(`/customers/${customerId}/credit-check/${creditCheckId}/analyze`);
-      return data;
-    },
-    onSuccess: () => {
-      toast.success('วิเคราะห์เครดิตเสร็จสิ้น');
-      queryClient.invalidateQueries({ queryKey: ['credit-checks'] });
-    },
-    onError: (err: unknown) => toast.error(getErrorMessage(err)),
-  });
-
-  const overrideMutation = useMutation({
-    mutationFn: async () => {
-      if (!overrideId || !overrideCustomerId) return;
-      const { data } = await api.post(`/customers/${overrideCustomerId}/credit-check/${overrideId}/override`, {
-        status: overrideStatus,
-        reviewNotes: overrideNotes || undefined,
-      });
-      return data;
-    },
-    onSuccess: () => {
-      toast.success('อัปเดตสถานะเครดิตเช็คแล้ว');
-      queryClient.invalidateQueries({ queryKey: ['credit-checks'] });
-      setOverrideId(null);
-      setOverrideCustomerId(null);
-      setOverrideStatus('');
-      setOverrideNotes('');
-    },
-    onError: (err: unknown) => toast.error(getErrorMessage(err)),
-  });
-
-  const handleBookBankScan = async (e: React.ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files?.[0];
-    if (!file) return;
-    if (bookBankFileRef.current) bookBankFileRef.current.value = '';
-    if (file.size > 10 * 1024 * 1024) {
-      toast.error('ไฟล์ต้องมีขนาดไม่เกิน 10MB');
-      return;
-    }
-    if (!file.type.startsWith('image/')) {
-      toast.error('กรุณาเลือกไฟล์รูปภาพ');
-      return;
-    }
-
-    setBookBankLoading(true);
-    try {
-      const imageBase64 = await compressImageForOcr(file);
-      const { data } = await api.post<OcrBookBankResult>('/ocr/book-bank', { imageBase64 }, { timeout: 90000 });
-      setBookBankResult(data);
-
-      // Auto-fill bank name from OCR
-      if (data.bankName) {
-        setBankName(data.bankName);
-      }
-
-      const pct = (data.confidence * 100).toFixed(0);
-      if (data.confidence < 0.7) {
-        toast.warning(`อ่านสมุดบัญชีสำเร็จ ความมั่นใจ ${pct}%`);
-      } else {
-        toast.success(`อ่านสมุดบัญชีสำเร็จ (ความมั่นใจ ${pct}%)`);
-      }
-    } catch (err: unknown) {
-      const axiosErr = err as { code?: string; response?: unknown };
-      if (axiosErr.code === 'ECONNABORTED' || !axiosErr.response) {
-        toast.error('ไม่สามารถเชื่อมต่อ OCR ได้ กรุณาลองใหม่');
-      } else {
-        toast.error(getErrorMessage(err));
-      }
-    } finally {
-      setBookBankLoading(false);
-    }
-  };
-
-  // Excel export
-  const exportExcel = async () => {
-    try {
-      toast.loading('กำลังสร้างไฟล์ Excel...', { id: 'excel-export' });
-      const exportParams = new URLSearchParams();
-      if (debouncedSearch) exportParams.set('search', debouncedSearch);
-      if (statusFilter) exportParams.set('status', statusFilter);
-      if (startDate) exportParams.set('startDate', startDate);
-      if (endDate) exportParams.set('endDate', endDate);
-      if (branchFilter) exportParams.set('branchId', branchFilter);
-      exportParams.set('page', '1');
-      exportParams.set('limit', '10000');
-      const { data: allData } = await api.get<CreditChecksResponse>(`/credit-checks?${exportParams}`);
-
-      const cols: ExcelColumn[] = [
-        { header: 'ชื่อลูกค้า', key: 'customerName', width: 25 },
-        { header: 'เบอร์โทร', key: 'phone', width: 14 },
-        { header: 'ธนาคาร', key: 'bankName', width: 16 },
-        { header: 'จำนวนเดือน Statement', key: 'statementMonths', width: 12 },
-        { header: 'คะแนน AI', key: 'aiScore', width: 10 },
-        { header: 'ความเสี่ยง', key: 'risk', width: 16 },
-        { header: 'สถานะ', key: 'status', width: 14 },
-      ];
-
-      if (isOwnerOrManager) {
-        cols.push(
-          { header: 'รายได้เฉลี่ย', key: 'monthlyIncome', width: 16 },
-          { header: 'ยอดเงินเฉลี่ย', key: 'averageBalance', width: 16 },
-        );
-      }
-
-      cols.push(
-        { header: 'อัตราภาระหนี้ (%)', key: 'affordabilityRatio', width: 14 },
-        { header: 'ความสม่ำเสมอรายได้', key: 'incomeConsistency', width: 16 },
-        { header: 'ผู้ตรวจ', key: 'checkedByName', width: 18 },
-        { header: 'หมายเหตุ', key: 'reviewNotes', width: 25 },
-        { header: 'เลขสัญญา', key: 'contractNumber', width: 18 },
-        { header: 'วันที่สร้าง', key: 'createdAt', width: 16 },
-      );
-
-      const getRiskLabel = (score: number | null) => {
-        if (score === null) return 'รอวิเคราะห์';
-        if (score >= 70) return 'ความเสี่ยงต่ำ';
-        if (score >= 50) return 'ความเสี่ยงปานกลาง';
-        if (score >= 40) return 'ต้องตรวจเพิ่ม';
-        return 'ความเสี่ยงสูง';
-      };
-
-      const now = new Date();
-      await exportToExcel({
-        columns: cols,
-        data: allData.data.map((cc: CreditCheckItem) => {
-          const ai = cc.aiAnalysis as AiAnalysisData | null;
-          const row: Record<string, unknown> = {
-            customerName: cc.customer.name,
-            phone: cc.customer.phone,
-            bankName: cc.bankName || '-',
-            statementMonths: cc.statementMonths,
-            aiScore: cc.aiScore ?? '-',
-            risk: getRiskLabel(cc.aiScore),
-            status: statusLabels[cc.status]?.label || cc.status,
-            affordabilityRatio: ai?.affordabilityRatio != null ? Number((ai.affordabilityRatio * 100).toFixed(1)) : '-',
-            incomeConsistency: ai?.incomeConsistency || '-',
-            checkedByName: cc.checkedBy?.name || '-',
-            reviewNotes: cc.reviewNotes || '-',
-            contractNumber: cc.contract?.contractNumber || '-',
-            createdAt: formatDateShort(cc.createdAt),
-          };
-          if (isOwnerOrManager) {
-            row.monthlyIncome = ai?.monthlyIncome != null ? Number(ai.monthlyIncome) : '-';
-            row.averageBalance = ai?.averageBalance != null ? Number(ai.averageBalance) : '-';
-          }
-          return row;
-        }),
-        sheetName: 'ตรวจสอบเครดิต',
-        filename: `ตรวจสอบเครดิต_${now.getFullYear()}-${String(now.getMonth() + 1).padStart(2, '0')}-${String(now.getDate()).padStart(2, '0')}.xlsx`,
-      });
-      toast.success(`ดาวน์โหลดสำเร็จ (${allData.data.length} รายการ)`, { id: 'excel-export' });
-    } catch {
-      toast.error('ไม่สามารถสร้างไฟล์ Excel ได้', { id: 'excel-export' });
-    }
-  };
-
-  const columns = [
-    {
-      key: 'customer',
-      label: 'ลูกค้า',
-      render: (cc: CreditCheckItem) => (
-        <div>
-          <button onClick={() => navigate(`/customers/${cc.customer.id}`)} className="text-sm font-medium text-primary hover:underline">{cc.customer.name}</button>
-          <div className="text-xs text-muted-foreground">{cc.customer.phone}</div>
-        </div>
-      ),
-    },
-    {
-      key: 'status',
-      label: 'สถานะ',
-      render: (cc: CreditCheckItem) => {
-        const s = statusLabels[cc.status] || { label: cc.status, className: 'bg-muted' };
-        return <span className={`px-2 py-0.5 rounded-full text-xs font-medium ${s.className}`}>{s.label}</span>;
-      },
-    },
-    {
-      key: 'aiScore',
-      label: 'คะแนน',
-      render: (cc: CreditCheckItem) => cc.aiScore !== null ? (
-        <div className="flex items-center gap-2">
-          <span className={`text-sm font-bold ${cc.aiScore >= 70 ? 'text-success' : cc.aiScore >= 50 ? 'text-amber-600' : 'text-destructive'}`}>{cc.aiScore}</span>
-          <div className="w-16 bg-muted rounded-full h-1.5">
-            <div className={`h-1.5 rounded-full ${cc.aiScore >= 70 ? 'bg-green-500' : cc.aiScore >= 50 ? 'bg-amber-500' : 'bg-red-500'}`} style={{ width: `${cc.aiScore}%` }} />
-          </div>
-        </div>
-      ) : <span className="text-xs text-muted-foreground">-</span>,
-    },
-    {
-      key: 'risk',
-      label: 'ความเสี่ยง',
-      render: (cc: CreditCheckItem) => getRiskBadge(cc.aiScore),
-    },
-    {
-      key: 'bankName',
-      label: 'ธนาคาร',
-      render: (cc: CreditCheckItem) => <span className="text-sm">{cc.bankName || '-'}</span>,
-    },
-    {
-      key: 'contract',
-      label: 'สัญญา',
-      render: (cc: CreditCheckItem) => cc.contract ? (
-        <button onClick={() => navigate(`/contracts/${cc.contract!.id}`)} className="text-xs text-primary hover:underline font-mono">{cc.contract.contractNumber}</button>
-      ) : <span className="text-xs text-muted-foreground">ยังไม่มีสัญญา</span>,
-    },
-    {
-      key: 'createdAt',
-      label: 'วันที่',
-      render: (cc: CreditCheckItem) => <span className="text-xs text-muted-foreground">{formatDateShort(cc.createdAt)}</span>,
-    },
-    {
-      key: 'actions',
-      label: '',
-      render: (cc: CreditCheckItem) => (
-        <div className="flex gap-2">
-          {cc.status === 'PENDING' && (
-            <button
-              onClick={() => analyzeMutation.mutate({ customerId: cc.customer.id, creditCheckId: cc.id })}
-              disabled={analyzeMutation.isPending}
-              className="px-3 py-1 text-xs bg-primary text-primary-foreground rounded-lg hover:bg-primary/90 disabled:opacity-50"
-            >
-              AI วิเคราะห์
-            </button>
-          )}
-          {canOverride && cc.aiScore !== null && (
-            <button
-              onClick={() => { setOverrideId(cc.id); setOverrideCustomerId(cc.customer.id); }}
-              className="px-3 py-1 text-xs bg-primary-100 text-primary-700 rounded-lg hover:bg-primary-200"
-            >
-              ปรับแก้
-            </button>
-          )}
-        </div>
-      ),
-    },
-  ];
-
-  const avgScoreColor = (score: number) => {
-    if (score >= 60) return 'text-success';
-    if (score >= 40) return 'text-amber-700 dark:text-amber-500';
-    return 'text-destructive';
-  };
-  const avgScoreBg = (score: number) => {
-    if (score >= 60) return 'bg-success/5 dark:bg-success/10 border-success/20';
-    if (score >= 40) return 'bg-amber-50 dark:bg-amber-500/10 border-amber-200 dark:border-amber-500/20';
-    return 'bg-destructive/5 dark:bg-destructive/10 border-destructive/20';
-  };
-  const avgScoreLabel = (score: number) => {
-    if (score >= 60) return 'text-success';
-    if (score >= 40) return 'text-amber-600 dark:text-amber-500';
-    return 'text-destructive';
-  };
-
-  return (
-    <div>
-      <PageHeader
-        title="ตรวจเครดิต"
-        subtitle="ตรวจสอบเครดิตลูกค้าก่อนทำสัญญา"
-        action={
-          <div className="flex gap-2">
-            <button
-              onClick={exportExcel}
-              className="px-4 py-2 text-sm border border-input rounded-lg hover:bg-muted"
-            >
-              ส่งออก Excel
-            </button>
-            <button
-              onClick={() => setShowCreateModal(true)}
-              className="px-4 py-2 text-sm bg-primary text-primary-foreground rounded-lg hover:bg-primary/90"
-            >
-              + ตรวจเครดิตใหม่
-            </button>
-          </div>
-        }
-      />
-
-      {/* Filters */}
-      <div className="flex gap-3 mb-4 flex-wrap items-end">
-        <input
-          type="text"
-          placeholder="ค้นหาชื่อลูกค้า..."
-          value={search}
-          onChange={(e) => setSearch(e.target.value)}
-          className="px-3 py-2 border border-input rounded-lg text-sm w-64"
-        />
-        <select
-          value={statusFilter}
-          onChange={(e) => setStatusFilter(e.target.value)}
-          className="px-3 py-2 border border-input rounded-lg text-sm"
-        >
-          <option value="">ทุกสถานะ</option>
-          <option value="PENDING">รอวิเคราะห์</option>
-          <option value="APPROVED">ผ่าน</option>
-          <option value="REJECTED">ไม่ผ่าน</option>
-          <option value="MANUAL_REVIEW">ต้องตรวจเพิ่ม</option>
-        </select>
-
-        {/* Date range */}
-        <div className="flex items-center gap-2">
-          <ThaiDateInput
-            value={startDate}
-            onChange={(e) => setStartDate(e.target.value)}
-            className="px-3 py-2 border border-input rounded-lg text-sm"
-          />
-          <span className="text-sm text-muted-foreground">ถึง</span>
-          <ThaiDateInput
-            value={endDate}
-            onChange={(e) => setEndDate(e.target.value)}
-            className="px-3 py-2 border border-input rounded-lg text-sm"
-          />
-        </div>
-        <div className="flex gap-1">
-          <button onClick={() => setDateRange('today')} className="px-2 py-1.5 text-xs border border-input rounded-lg hover:bg-muted">วันนี้</button>
-          <button onClick={() => setDateRange('week')} className="px-2 py-1.5 text-xs border border-input rounded-lg hover:bg-muted">สัปดาห์นี้</button>
-          <button onClick={() => setDateRange('month')} className="px-2 py-1.5 text-xs border border-input rounded-lg hover:bg-muted">เดือนนี้</button>
-          {(startDate || endDate) && (
-            <button onClick={() => { setStartDate(''); setEndDate(''); }} className="px-2 py-1.5 text-xs text-destructive border border-destructive/20 rounded-lg hover:bg-destructive/5">ล้าง</button>
-          )}
-        </div>
-
-        {/* Branch filter (OWNER only) */}
-        {isOwner && branches.length > 0 && (
-          <select
-            value={branchFilter}
-            onChange={(e) => setBranchFilter(e.target.value)}
-            className="px-3 py-2 border border-input rounded-lg text-sm"
-          >
-            <option value="">ทุกสาขา</option>
-            {branches.map((b) => (
-              <option key={b.id} value={b.id}>{b.name}</option>
-            ))}
-          </select>
-        )}
-      </div>
-
-      {/* Summary cards */}
-      <div className="grid grid-cols-2 md:grid-cols-5 gap-5 lg:gap-7.5 mb-6">
-        <div className="bg-card rounded-lg border border-l-[3px] border-l-foreground p-4 hover:shadow-card-hover transition-all">
-          <div className="text-xs text-muted-foreground">ทั้งหมด</div>
-          <div className="text-xl font-bold">{summary?.totalCount ?? 0}</div>
-        </div>
-        <div className="bg-success/5 dark:bg-success/10 rounded-lg border border-success/20 border-l-[3px] border-l-success p-4 hover:shadow-card-hover transition-all">
-          <div className="text-xs text-success">ผ่าน</div>
-          <div className="text-xl font-bold text-success">{summary?.approvedCount ?? 0}</div>
-        </div>
-        <div className="bg-amber-50 dark:bg-amber-500/10 rounded-lg border border-amber-200 dark:border-amber-500/20 border-l-[3px] border-l-warning p-4 hover:shadow-card-hover transition-all">
-          <div className="text-xs text-amber-600 dark:text-amber-500">รอวิเคราะห์ / ตรวจเพิ่ม</div>
-          <div className="text-xl font-bold text-amber-700 dark:text-amber-500">{summary?.pendingCount ?? 0}</div>
-        </div>
-        <div className="bg-destructive/5 dark:bg-destructive/10 rounded-lg border border-destructive/20 border-l-[3px] border-l-destructive p-4 hover:shadow-card-hover transition-all">
-          <div className="text-xs text-destructive">ไม่ผ่าน</div>
-          <div className="text-xl font-bold text-destructive">{summary?.rejectedCount ?? 0}</div>
-        </div>
-        <div className={`rounded-lg border p-4 hover:shadow-card-hover transition-all ${summary?.avgScore ? avgScoreBg(summary.avgScore) : 'bg-card'}`}>
-          <div className={`text-xs ${summary?.avgScore ? avgScoreLabel(summary.avgScore) : 'text-muted-foreground'}`}>คะแนน AI เฉลี่ย</div>
-          <div className={`text-xl font-bold ${summary?.avgScore ? avgScoreColor(summary.avgScore) : ''}`}>
-            {summary?.avgScore ?? '-'}
-          </div>
-        </div>
-      </div>
-
-      <QueryBoundary
-        isLoading={isLoading}
-        isError={isError}
-        error={error}
-        onRetry={refetch}
-        errorTitle="ไม่สามารถโหลดรายการตรวจเครดิตได้"
-      >
-        <>
-          <DataTable
-            columns={columns}
-            data={creditChecks}
-            emptyMessage="ยังไม่มีรายการตรวจเครดิต"
-            onRowClick={(cc: CreditCheckItem) => setExpandedRow(expandedRow === cc.id ? null : cc.id)}
-            pagination={creditChecksData ? {
-              page: creditChecksData.page,
-              totalPages: creditChecksData.totalPages,
-              total: creditChecksData.total,
-              onPageChange: setPage,
-            } : undefined}
-          />
-
-          {/* Expanded AI detail */}
-          {expandedRow && creditChecks.find((cc) => cc.id === expandedRow) && (() => {
-            const cc = creditChecks.find((c) => c.id === expandedRow)!;
-            const ai = cc.aiAnalysis as AiAnalysisData | null;
-            return (
-              <div className="mt-2 mb-4 bg-muted/50 border rounded-lg p-4 space-y-3">
-                <div className="flex items-center justify-between">
-                  <h4 className="text-sm font-semibold">รายละเอียดการวิเคราะห์ AI — {cc.customer.name}</h4>
-                  <button onClick={() => setExpandedRow(null)} className="text-xs text-muted-foreground hover:text-foreground">ปิด</button>
-                </div>
-
-                {ai ? (
-                  <div className="grid grid-cols-2 md:grid-cols-4 gap-3">
-                    {ai.monthlyIncome != null && (
-                      <div className="bg-card rounded border p-3">
-                        <div className="text-xs text-muted-foreground">รายได้เฉลี่ย/เดือน</div>
-                        <div className="text-sm font-bold">{Number(ai.monthlyIncome).toLocaleString()} ฿</div>
-                      </div>
-                    )}
-                    {ai.averageBalance != null && (
-                      <div className="bg-card rounded border p-3">
-                        <div className="text-xs text-muted-foreground">ยอดเงินเฉลี่ย</div>
-                        <div className="text-sm font-bold">{Number(ai.averageBalance).toLocaleString()} ฿</div>
-                      </div>
-                    )}
-                    {ai.affordabilityRatio != null && (
-                      <div className="bg-card rounded border p-3">
-                        <div className="text-xs text-muted-foreground">อัตราภาระหนี้</div>
-                        <div className="text-sm font-bold">{((ai.affordabilityRatio ?? 0) * 100).toFixed(1)}%</div>
-                      </div>
-                    )}
-                    {ai.incomeConsistency && (
-                      <div className="bg-card rounded border p-3">
-                        <div className="text-xs text-muted-foreground">ความสม่ำเสมอรายได้</div>
-                        <div className={`text-sm font-bold ${ai.incomeConsistency === 'stable' ? 'text-success' : 'text-amber-600'}`}>
-                          {ai.incomeConsistency === 'stable' ? 'สม่ำเสมอ' : 'ไม่สม่ำเสมอ'}
-                        </div>
-                      </div>
-                    )}
-                  </div>
-                ) : (
-                  <div className="text-sm text-muted-foreground">ยังไม่มีข้อมูลจากการวิเคราะห์ AI</div>
-                )}
-
-                {/* Risk & Positive factors */}
-                <div className="grid grid-cols-1 md:grid-cols-2 gap-3">
-                  {ai?.riskFactors && ai.riskFactors.length > 0 && (
-                    <div className="bg-destructive/5 dark:bg-destructive/10 border border-destructive/20 rounded p-3">
-                      <div className="text-xs font-medium text-destructive mb-1">ปัจจัยเสี่ยง</div>
-                      <ul className="space-y-1">
-                        {ai.riskFactors.map((f, i) => (
-                          <li key={i} className="text-xs text-destructive flex items-start gap-1">
-                            <span className="mt-0.5">•</span> {f}
-                          </li>
-                        ))}
-                      </ul>
-                    </div>
-                  )}
-                  {ai?.positiveFactors && ai.positiveFactors.length > 0 && (
-                    <div className="bg-success/5 dark:bg-success/10 border border-success/20 rounded p-3">
-                      <div className="text-xs font-medium text-success mb-1">ปัจจัยบวก</div>
-                      <ul className="space-y-1">
-                        {ai.positiveFactors.map((f, i) => (
-                          <li key={i} className="text-xs text-success flex items-start gap-1">
-                            <span className="mt-0.5">•</span> {f}
-                          </li>
-                        ))}
-                      </ul>
-                    </div>
-                  )}
-                </div>
-
-                {/* AI Summary & Recommendation */}
-                {(cc.aiSummary || cc.aiRecommendation) && (
-                  <div className="space-y-2">
-                    {cc.aiSummary && (
-                      <div className="bg-card rounded border p-3">
-                        <div className="text-2xs font-medium text-muted-foreground uppercase tracking-wider mb-2">สรุปผลวิเคราะห์</div>
-                        <div className="text-sm">{cc.aiSummary}</div>
-                      </div>
-                    )}
-                    {cc.aiRecommendation && (
-                      <div className="bg-card rounded border p-3">
-                        <div className="text-2xs font-medium text-muted-foreground uppercase tracking-wider mb-2">คำแนะนำ</div>
-                        <div className="text-sm">{cc.aiRecommendation}</div>
-                      </div>
-                    )}
-                  </div>
-                )}
-              </div>
-            );
-          })()}
-        </>
-      </QueryBoundary>
-
-      {/* Create Full-Screen Overlay */}
-      {showCreateModal && (
-        <div className="fixed inset-0 z-50 bg-black/50 backdrop-blur-sm flex items-start justify-center pt-8 pb-8">
-          <div className="w-full max-w-4xl bg-background rounded-xl shadow-2xl overflow-y-auto max-h-[calc(100vh-4rem)]">
-            {/* Header */}
-            <div className="sticky top-0 z-10 bg-background/95 backdrop-blur-sm border-b px-6 py-4 flex items-center justify-between">
-              <button onClick={() => { setShowCreateModal(false); setSelectedCustomer(null); setBankName(''); setCustomerSearch(''); }} className="flex items-center gap-1.5 text-sm text-muted-foreground hover:text-foreground transition-colors">
-                <ArrowLeft className="size-4" /> กลับ
-              </button>
-              <h2 className="text-lg font-semibold text-foreground">ตรวจเครดิตใหม่</h2>
-              <div className="w-16" />
-            </div>
-
-            <div className="p-6 space-y-5">
-
-              {/* ─── Section 1: ข้อมูลลูกค้า ─── */}
-              <div className="rounded-xl border border-border bg-card p-5">
-                <div className="flex items-center gap-2.5 mb-4">
-                  <div className="flex items-center justify-center size-8 rounded-lg bg-primary/10 text-primary">
-                    <User className="size-4" strokeWidth={1.5} />
-                  </div>
-                  <div>
-                    <h3 className="text-sm font-semibold text-foreground">ข้อมูลลูกค้า</h3>
-                    <p className="text-xs text-muted-foreground">เลือกลูกค้าที่ต้องการตรวจเครดิต</p>
-                  </div>
-                </div>
-
-                {!selectedCustomer ? (
-                  <div>
-                    <input
-                      type="text"
-                      placeholder="ค้นหาชื่อ, เบอร์โทร, เลขบัตร..."
-                      value={customerSearch}
-                      onChange={(e) => setCustomerSearch(e.target.value)}
-                      className="w-full px-3 py-2 border border-input rounded-lg text-sm mb-3"
-                    />
-                    <div className="max-h-60 overflow-y-auto space-y-2">
-                      {customers.map((c) => (
-                        <div
-                          key={c.id}
-                          onClick={() => setSelectedCustomer(c)}
-                          className="p-3 rounded-lg border cursor-pointer hover:border-primary/40 hover:bg-primary/5 dark:hover:bg-primary/10 transition-colors"
-                        >
-                          <div className="text-sm font-medium">{c.name}</div>
-                          <div className="text-xs text-muted-foreground">{c.phone} {c.salary ? `| เงินเดือน ${parseFloat(c.salary).toLocaleString()} ฿` : ''}</div>
-                        </div>
-                      ))}
-                      {customers.length === 0 && customerSearch && (
-                        <div className="text-center py-4 text-sm text-muted-foreground">ไม่พบลูกค้า</div>
-                      )}
-                    </div>
-                  </div>
-                ) : (
-                  <div>
-                    <div className="bg-primary/5 dark:bg-primary/10 rounded-xl p-4 flex items-center justify-between">
-                      <div className="flex-1">
-                        <div className="flex items-center gap-2 mb-1">
-                          <span className="text-sm font-semibold text-primary">{selectedCustomer.name}</span>
-                          {customerHistory ? (
-                            customerHistory.isReturning
-                              ? <span className="px-2 py-0.5 rounded-full text-xs font-medium bg-success/10 text-success">ลูกค้าเก่า</span>
-                              : <span className="px-2 py-0.5 rounded-full text-xs font-medium bg-amber-100 text-amber-700">ลูกค้าใหม่</span>
-                          ) : null}
-                        </div>
-                        <div className="grid grid-cols-2 md:grid-cols-4 gap-3 mt-2">
-                          <div>
-                            <div className="text-2xs text-muted-foreground">เบอร์โทร</div>
-                            <div className="text-xs font-medium">{selectedCustomer.phone}</div>
-                          </div>
-                          <div>
-                            <div className="text-2xs text-muted-foreground">อาชีพ</div>
-                            <div className="text-xs font-medium">{selectedCustomer.occupation || '-'}</div>
-                          </div>
-                          <div>
-                            <div className="text-2xs text-muted-foreground">เงินเดือน</div>
-                            <div className="text-xs font-medium">{selectedCustomer.salary ? `${parseFloat(selectedCustomer.salary).toLocaleString()} ฿` : '-'}</div>
-                          </div>
-                          <div>
-                            <div className="text-2xs text-muted-foreground">ประเภทที่อยู่</div>
-                            <div className="text-xs font-medium">{selectedCustomer.addressCurrentType || '-'}</div>
-                          </div>
-                        </div>
-                      </div>
-                      <button onClick={() => setSelectedCustomer(null)} className="text-xs text-primary hover:text-primary/80 ml-4">เปลี่ยน</button>
-                    </div>
-                  </div>
-                )}
-              </div>
-
-              {/* ─── Section 2: เอกสารประกอบ ─── */}
-              {selectedCustomer && (
-                <div className="rounded-xl border border-border bg-card p-5">
-                  <div className="flex items-center gap-2.5 mb-4">
-                    <div className="flex items-center justify-center size-8 rounded-lg bg-emerald-500/10 text-emerald-500">
-                      <FileText className="size-4" strokeWidth={1.5} />
-                    </div>
-                    <div>
-                      <h3 className="text-sm font-semibold text-foreground">เอกสารประกอบ</h3>
-                      <p className="text-xs text-muted-foreground">สลิปเงินเดือนและ Statement ธนาคาร</p>
-                    </div>
-                  </div>
-
-                  {/* Part A: สลิปเงินเดือน */}
-                  <div className="mb-5 p-4 bg-muted/30 rounded-lg border border-border">
-                    <h4 className="text-xs font-semibold text-foreground mb-3 uppercase tracking-wider">A. สลิปเงินเดือน</h4>
-                    <div className="flex items-center gap-3 mb-3">
-                      <input
-                        ref={salarySlipFileRef}
-                        type="file"
-                        accept="image/*"
-                        multiple
-                        onChange={(e) => {
-                          if (e.target.files) {
-                            const files = Array.from(e.target.files).slice(0, 3);
-                            setSalarySlipFiles(files);
-                          }
-                        }}
-                        className="flex-1 text-sm text-muted-foreground file:mr-3 file:py-2 file:px-4 file:rounded-lg file:border-0 file:text-sm file:font-medium file:bg-emerald-500/10 file:text-emerald-600"
-                      />
-                      <button
-                        onClick={handleSalarySlipOcr}
-                        disabled={salarySlipLoading || salarySlipFiles.length === 0}
-                        className="inline-flex items-center gap-2 px-4 py-2 bg-emerald-600 text-white rounded-lg text-xs font-medium hover:bg-emerald-700 disabled:opacity-50"
-                      >
-                        {salarySlipLoading ? (
-                          <>
-                            <div className="animate-spin rounded-full h-3 w-3 border-b-2 border-white" />
-                            กำลังวิเคราะห์...
-                          </>
-                        ) : 'AI วิเคราะห์'}
-                      </button>
-                    </div>
-                    <div className="text-2xs text-muted-foreground mb-2">รองรับสูงสุด 3 รูป</div>
-
-                    {salarySlipResult && (
-                      <div className="grid grid-cols-2 md:grid-cols-4 gap-3 mt-3">
-                        <div>
-                          <label className="block text-2xs text-muted-foreground mb-1">เงินเดือนสุทธิ</label>
-                          <input
-                            type="text"
-                            value={salarySlipEditable.netSalary}
-                            onChange={(e) => setSalarySlipEditable(prev => ({ ...prev, netSalary: e.target.value }))}
-                            className="w-full px-2 py-1.5 border border-input rounded-lg text-sm"
-                          />
-                        </div>
-                        <div>
-                          <label className="block text-2xs text-muted-foreground mb-1">ชื่อบริษัท</label>
-                          <input
-                            type="text"
-                            value={salarySlipEditable.employerName}
-                            onChange={(e) => setSalarySlipEditable(prev => ({ ...prev, employerName: e.target.value }))}
-                            className="w-full px-2 py-1.5 border border-input rounded-lg text-sm"
-                          />
-                        </div>
-                        <div>
-                          <label className="block text-2xs text-muted-foreground mb-1">วันเงินเดือนออก</label>
-                          <input
-                            type="text"
-                            value={salarySlipEditable.payDay}
-                            onChange={(e) => setSalarySlipEditable(prev => ({ ...prev, payDay: e.target.value }))}
-                            className="w-full px-2 py-1.5 border border-input rounded-lg text-sm"
-                            placeholder="เช่น 25"
-                          />
-                        </div>
-                        <div>
-                          <label className="block text-2xs text-muted-foreground mb-1">ธนาคาร</label>
-                          <input
-                            type="text"
-                            value={salarySlipEditable.bankName}
-                            onChange={(e) => setSalarySlipEditable(prev => ({ ...prev, bankName: e.target.value }))}
-                            className="w-full px-2 py-1.5 border border-input rounded-lg text-sm"
-                          />
-                        </div>
-                      </div>
-                    )}
-                  </div>
-
-                  {/* Part B: Statement ธนาคาร */}
-                  <div className="p-4 bg-muted/30 rounded-lg border border-border">
-                    <h4 className="text-xs font-semibold text-foreground mb-3 uppercase tracking-wider">B. Statement ธนาคาร</h4>
-                    <div className="mb-3">
-                      <label className="block text-2xs text-muted-foreground mb-1">เลือกธนาคาร</label>
-                      <select
-                        value={statementBankName}
-                        onChange={(e) => setStatementBankName(e.target.value)}
-                        className="w-full px-3 py-2 border border-input rounded-lg text-sm"
-                      >
-                        <option value="">-- เลือกธนาคาร --</option>
-                        {BANK_OPTIONS.map((b) => <option key={b} value={b}>{b}</option>)}
-                      </select>
-                    </div>
-                    <div className="flex items-center gap-3 mb-3">
-                      <input
-                        ref={statementFileRef}
-                        type="file"
-                        accept="image/*,.pdf"
-                        multiple
-                        onChange={(e) => {
-                          if (e.target.files) setStatementFiles(Array.from(e.target.files));
-                        }}
-                        className="flex-1 text-sm text-muted-foreground file:mr-3 file:py-2 file:px-4 file:rounded-lg file:border-0 file:text-sm file:font-medium file:bg-emerald-500/10 file:text-emerald-600"
-                      />
-                      <button
-                        onClick={handleStatementOcr}
-                        disabled={statementLoading || statementFiles.length === 0}
-                        className="inline-flex items-center gap-2 px-4 py-2 bg-emerald-600 text-white rounded-lg text-xs font-medium hover:bg-emerald-700 disabled:opacity-50"
-                      >
-                        {statementLoading ? (
-                          <>
-                            <div className="animate-spin rounded-full h-3 w-3 border-b-2 border-white" />
-                            กำลังวิเคราะห์...
-                          </>
-                        ) : 'AI วิเคราะห์'}
-                      </button>
-                    </div>
-
-                    {statementResult && (
-                      <div className="grid grid-cols-3 gap-3 mt-3">
-                        <div className="bg-card rounded border p-3">
-                          <div className="text-2xs text-muted-foreground">ยอดเข้าเฉลี่ย</div>
-                          <div className="text-sm font-bold text-success">{statementResult.totalIncome != null ? `${statementResult.totalIncome.toLocaleString()} ฿` : '-'}</div>
-                        </div>
-                        <div className="bg-card rounded border p-3">
-                          <div className="text-2xs text-muted-foreground">ยอดออกเฉลี่ย</div>
-                          <div className="text-sm font-bold text-destructive">{statementResult.totalExpense != null ? `${statementResult.totalExpense.toLocaleString()} ฿` : '-'}</div>
-                        </div>
-                        <div className="bg-card rounded border p-3">
-                          <div className="text-2xs text-muted-foreground">ยอดคงเหลือ</div>
-                          <div className="text-sm font-bold">{statementResult.balance != null ? `${statementResult.balance.toLocaleString()} ฿` : '-'}</div>
-                        </div>
-                      </div>
-                    )}
-                  </div>
-                </div>
-              )}
-
-              {/* ─── Section 3: ประวัติในระบบ ─── */}
-              {selectedCustomer && (
-                <div className="rounded-xl border border-border bg-card p-5">
-                  <div className="flex items-center gap-2.5 mb-4">
-                    <div className="flex items-center justify-center size-8 rounded-lg bg-violet-500/10 text-violet-500">
-                      <Clock className="size-4" strokeWidth={1.5} />
-                    </div>
-                    <div>
-                      <h3 className="text-sm font-semibold text-foreground">ประวัติในระบบ</h3>
-                      <p className="text-xs text-muted-foreground">ข้อมูลสัญญาและประวัติชำระเงินจากระบบ</p>
-                    </div>
-                    {customerHistory && (
-                      <div className="ml-auto">
-                        {customerHistory.isReturning
-                          ? <span className="px-2.5 py-1 rounded-full text-xs font-medium bg-success/10 text-success">ลูกค้าเก่า</span>
-                          : <span className="px-2.5 py-1 rounded-full text-xs font-medium bg-amber-100 text-amber-700">ลูกค้าใหม่</span>
-                        }
-                      </div>
-                    )}
-                  </div>
-
-                  {customerHistory ? (
-                    <div className="grid grid-cols-2 md:grid-cols-4 gap-3">
-                      <div className="bg-muted/50 rounded-lg border p-3">
-                        <div className="text-2xs text-muted-foreground">สัญญาทั้งหมด</div>
-                        <div className="text-lg font-bold">{customerHistory.totalContracts} <span className="text-xs font-normal text-muted-foreground">สัญญา</span></div>
-                      </div>
-                      <div className="bg-muted/50 rounded-lg border p-3">
-                        <div className="text-2xs text-muted-foreground">ปิดแล้ว / ค้างอยู่</div>
-                        <div className="text-lg font-bold">
-                          <span className="text-success">{customerHistory.closedContracts}</span>
-                          <span className="text-xs font-normal text-muted-foreground"> / </span>
-                          <span className="text-amber-600">{customerHistory.activeContracts}</span>
-                        </div>
-                      </div>
-                      <div className="bg-muted/50 rounded-lg border p-3">
-                        <div className="text-2xs text-muted-foreground">ชำระตรงเวลา</div>
-                        <div className={`text-lg font-bold ${(customerHistory.onTimePaymentPct ?? 0) >= 80 ? 'text-success' : (customerHistory.onTimePaymentPct ?? 0) >= 50 ? 'text-amber-600' : 'text-destructive'}`}>
-                          {(customerHistory.onTimePaymentPct ?? 0).toFixed(0)}%
-                        </div>
-                      </div>
-                      <div className="bg-muted/50 rounded-lg border p-3">
-                        <div className="text-2xs text-muted-foreground">ยอดค้างปัจจุบัน</div>
-                        <div className="text-lg font-bold">{customerHistory.currentOutstanding.toLocaleString()} <span className="text-xs font-normal text-muted-foreground">฿</span></div>
-                      </div>
-                    </div>
-                  ) : (
-                    <div className="text-sm text-muted-foreground text-center py-4">กำลังโหลดข้อมูล...</div>
-                  )}
-                </div>
-              )}
-
-              {/* ─── Section 4: สรุปวิเคราะห์ ─── */}
-              {selectedCustomer && (
-                <div className="rounded-xl border border-border bg-card p-5">
-                  <div className="flex items-center gap-2.5 mb-4">
-                    <div className="flex items-center justify-center size-8 rounded-lg bg-orange-500/10 text-orange-500">
-                      <BarChart3 className="size-4" strokeWidth={1.5} />
-                    </div>
-                    <div>
-                      <h3 className="text-sm font-semibold text-foreground">สรุปวิเคราะห์</h3>
-                      <p className="text-xs text-muted-foreground">คำนวณ Risk Score และคำแนะนำ</p>
-                    </div>
-                  </div>
-
-                  {!riskScore ? (
-                    <div className="text-center py-4">
-                      <button
-                        onClick={() => handleCalculateRisk(selectedCustomer.id)}
-                        disabled={riskLoading}
-                        className="inline-flex items-center gap-2 px-5 py-2.5 bg-orange-500 text-white rounded-lg text-sm font-medium hover:bg-orange-600 disabled:opacity-50"
-                      >
-                        {riskLoading ? (
-                          <>
-                            <div className="animate-spin rounded-full h-4 w-4 border-b-2 border-white" />
-                            กำลังคำนวณ...
-                          </>
-                        ) : 'คำนวณ Risk Score'}
-                      </button>
-                    </div>
-                  ) : (
-                    <div className="space-y-4">
-                      <div className="grid grid-cols-2 md:grid-cols-3 gap-3">
-                        <div className="bg-muted/50 rounded-lg border p-3">
-                          <div className="text-2xs text-muted-foreground mb-1">Risk Score</div>
-                          <span className={`px-3 py-1 rounded-full text-sm font-bold ${
-                            riskScore.riskLevel === 'LOW' ? 'bg-success/10 text-success' :
-                            riskScore.riskLevel === 'MEDIUM' ? 'bg-amber-100 text-amber-700' :
-                            'bg-destructive/10 text-destructive'
-                          }`}>
-                            {riskScore.riskLevel === 'LOW' ? 'ต่ำ' : riskScore.riskLevel === 'MEDIUM' ? 'ปานกลาง' : 'สูง'}
-                            {' '}({riskScore.score})
-                          </span>
-                        </div>
-                        <div className="bg-muted/50 rounded-lg border p-3">
-                          <div className="text-2xs text-muted-foreground mb-1">Debt-to-Income</div>
-                          <div className={`text-lg font-bold ${(riskScore.debtToIncome ?? 0) <= 40 ? 'text-success' : (riskScore.debtToIncome ?? 0) <= 60 ? 'text-amber-600' : 'text-destructive'}`}>
-                            {(riskScore.debtToIncome ?? 0).toFixed(1)}%
-                          </div>
-                          <div className="text-2xs text-muted-foreground">ค่างวด / เงินเดือน</div>
-                        </div>
-                        {riskScore.recommendedPayDay && (
-                          <div className="bg-muted/50 rounded-lg border p-3">
-                            <div className="text-2xs text-muted-foreground mb-1">แนะนำกำหนดชำระ</div>
-                            <div className="text-lg font-bold text-primary">วันที่ {riskScore.recommendedPayDay}</div>
-                            <div className="text-2xs text-muted-foreground">= วันเงินเดือนออก</div>
-                          </div>
-                        )}
-                      </div>
-
-                      {riskScore.recommendation && (
-                        <div className="bg-muted/30 rounded-lg border p-3">
-                          <div className="text-2xs font-medium text-muted-foreground uppercase tracking-wider mb-1">คำแนะนำ AI</div>
-                          <div className="text-sm">{riskScore.recommendation}</div>
-                        </div>
-                      )}
-                    </div>
-                  )}
-
-                  <div className="mt-4">
-                    <label className="block text-xs font-medium text-foreground mb-1.5">หมายเหตุ</label>
-                    <textarea
-                      value={reviewNotesDraft}
-                      onChange={(e) => setReviewNotesDraft(e.target.value)}
-                      rows={2}
-                      placeholder="ระบุหมายเหตุเพิ่มเติม (ถ้ามี)..."
-                      className="w-full px-3 py-2 border border-input rounded-lg text-sm"
-                    />
-                  </div>
-                </div>
-              )}
-            </div>
-
-            {/* Footer */}
-            {selectedCustomer && (
-              <div className="sticky bottom-0 z-10 bg-background/95 backdrop-blur-sm border-t px-6 py-4 flex items-center justify-between">
-                <button
-                  onClick={() => { setShowCreateModal(false); setSelectedCustomer(null); setBankName(''); setCustomerSearch(''); }}
-                  className="px-4 py-2 text-sm border border-input rounded-lg hover:bg-muted"
-                >
-                  ยกเลิก
-                </button>
-                <div className="flex gap-2">
-                  <button
-                    onClick={() => {
-                      if (!selectedCustomer) return;
-                      const fileUrls: string[] = [];
-                      // Simple save as draft: create with minimal data
-                      const files = fileRef.current?.files;
-                      if (files && files.length > 0) {
-                        uploadMutation.mutate(files);
-                      } else {
-                        // Create without statement files
-                        api.post(`/customers/${selectedCustomer.id}/credit-check`, {
-                          bankName: statementBankName || bankName || undefined,
-                          statementFiles: fileUrls,
-                          statementMonths: 3,
-                          reviewNotes: reviewNotesDraft || undefined,
-                        }).then(() => {
-                          toast.success('บันทึกร่างตรวจเครดิตสำเร็จ');
-                          queryClient.invalidateQueries({ queryKey: ['credit-checks'] });
-                          setShowCreateModal(false);
-                          setSelectedCustomer(null);
-                        }).catch((err: unknown) => toast.error(getErrorMessage(err)));
-                      }
-                    }}
-                    disabled={uploadMutation.isPending}
-                    className="px-4 py-2 text-sm border border-input rounded-lg hover:bg-muted"
-                  >
-                    บันทึก
-                  </button>
-                  <button
-                    onClick={() => {
-                      if (!selectedCustomer) return;
-                      api.post(`/customers/${selectedCustomer.id}/credit-check`, {
-                        bankName: statementBankName || bankName || undefined,
-                        statementFiles: [],
-                        statementMonths: 3,
-                        reviewNotes: reviewNotesDraft || undefined,
-                        status: 'APPROVED',
-                      }).then(() => {
-                        toast.success('อนุมัติเครดิตสำเร็จ');
-                        queryClient.invalidateQueries({ queryKey: ['credit-checks'] });
-                        setShowCreateModal(false);
-                        setSelectedCustomer(null);
-                      }).catch((err: unknown) => toast.error(getErrorMessage(err)));
-                    }}
-                    className="px-4 py-2 text-sm bg-green-600 text-white rounded-lg hover:bg-green-700"
-                  >
-                    อนุมัติ
-                  </button>
-                  <button
-                    onClick={() => {
-                      if (!selectedCustomer) return;
-                      api.post(`/customers/${selectedCustomer.id}/credit-check`, {
-                        bankName: statementBankName || bankName || undefined,
-                        statementFiles: [],
-                        statementMonths: 3,
-                        reviewNotes: reviewNotesDraft || undefined,
-                        status: 'REJECTED',
-                      }).then(() => {
-                        toast.success('ปฏิเสธเครดิตแล้ว');
-                        queryClient.invalidateQueries({ queryKey: ['credit-checks'] });
-                        setShowCreateModal(false);
-                        setSelectedCustomer(null);
-                      }).catch((err: unknown) => toast.error(getErrorMessage(err)));
-                    }}
-                    className="px-4 py-2 text-sm bg-red-600 text-white rounded-lg hover:bg-red-700"
-                  >
-                    ไม่อนุมัติ
-                  </button>
-                </div>
-              </div>
-            )}
-          </div>
-        </div>
-      )}
-
-      {/* Override Modal */}
-      {overrideId && (
-        <Modal isOpen title="ปรับแก้สถานะเครดิตเช็ค" onClose={() => { setOverrideId(null); setOverrideCustomerId(null); }}>
-          <div className="space-y-4">
-            <div>
-              <label className="block text-sm font-medium text-foreground mb-1">สถานะใหม่</label>
-              <select
-                value={overrideStatus}
-                onChange={(e) => setOverrideStatus(e.target.value)}
-                className="w-full px-3 py-2 border border-input rounded-lg text-sm"
-              >
-                <option value="">เลือกสถานะ...</option>
-                <option value="APPROVED">อนุมัติ</option>
-                <option value="REJECTED">ปฏิเสธ</option>
-                <option value="MANUAL_REVIEW">ตรวจเพิ่มเติม</option>
-              </select>
-            </div>
-            <div>
-              <label className="block text-sm font-medium text-foreground mb-1">หมายเหตุ</label>
-              <textarea
-                value={overrideNotes}
-                onChange={(e) => setOverrideNotes(e.target.value)}
-                rows={2}
-                placeholder="ระบุเหตุผล..."
-                className="w-full px-3 py-2 border border-input rounded-lg text-sm"
-              />
-            </div>
-            <div className="flex gap-3">
-              <button onClick={() => { setOverrideId(null); setOverrideCustomerId(null); }} className="flex-1 px-4 py-2 text-sm border border-input rounded-lg">ยกเลิก</button>
-              <button
-                onClick={() => overrideMutation.mutate()}
-                disabled={!overrideStatus || overrideMutation.isPending}
-                className="flex-1 px-4 py-2 text-sm bg-primary text-primary-foreground rounded-lg hover:bg-primary/90 disabled:opacity-50"
-              >
-                {overrideMutation.isPending ? 'กำลังบันทึก...' : 'บันทึก'}
-              </button>
-            </div>
-          </div>
-        </Modal>
-      )}
-    </div>
-  );
-}
