+++ apps/web/src/pages/ForgotPasswordPage.tsx
@@ -59,6 +59,7 @@ export default function ForgotPasswordPage() {
                 onChange={(e) => setEmail(e.target.value)}
                 className="w-full h-10 px-3.5 border border-input rounded-lg text-sm outline-none transition-all bg-background text-foreground placeholder:text-muted-foreground focus-visible:ring-2 focus-visible:ring-ring/30 focus-visible:ring-offset-[3px] focus-visible:ring-offset-background"
                 placeholder="email@example.com"
+                autoComplete="email"
                 required
               />
             </div>
