+++ apps/web/src/pages/SuppliersPage.tsx
@@ -1,4 +1,5 @@
 import { useState, useEffect } from 'react';
+import { useDocumentTitle } from '@/hooks/useDocumentTitle';
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
 import { useNavigate } from 'react-router-dom';
 import { toast } from 'sonner';
@@ -86,6 +87,7 @@ function formatTaxId(value: string): string {
 }
 
 export default function SuppliersPage() {
+  useDocumentTitle('ผู้ขาย');
   const { user } = useAuth();
   const navigate = useNavigate();
   const queryClient = useQueryClient();
